sfsfsd

+++ src/freequant/FreeQuant.Instruments/MarketDataType.cs
@@ -2,10 +2,10 @@ namespace FreeQuant.Instruments
 {
 	public enum MarketDataType : byte
 	{
-		None = (byte)0,
-		Trade = (byte)1,
-		Quote = (byte)2,
-		MarketDepth = (byte)4,
-		All = (byte)255,
+		None = 0,
+		Trade = 1,
+		Quote = 2,
+		MarketDepth = 4,
+		All = 255,
 	}
 }
